// packages/core/src/types.ts
export type Agent = {
  output: unknown
  speaker: string
  agent: string
  client: string
  channel: string
  entity: number
  roomInfo?: {
    user: string
    inConversation: boolean
    isBot: boolean
    info3d: string
  }[]
}

export interface IRunContextEditor extends NodeEditor {
  thoth: EditorContext
  abort: Function
}

export type DataSocketType = {
  name: SocketNameType
  taskType: 'output' | 'option'
  socketKey: string
  connectionType: 'input' | 'output'
  socketType: SocketType
  useSocketName: boolean
}

export type ThothNode = Node & {
  inspector: Inspector
  display: (content: string) => void
  outputs: { name: string; [key: string]: unknown }[]
  category?: string
  deprecated?: boolean
  displayName?: string
  info: string
  subscription: Function
  console: ThothConsole
}
// ...
export type NodeData = ReteNodeData & {
  fewshot?: string
  display: Function
  error?: boolean
  console: ThothConsole
}

// export type Node = {
//   id: number,
//   data: NodeData,
//   name: string,
//   inputs: NodeOutputs,
//   outputs?: NodeOutputs,
//   position: number[]
// }

// export type Spell = {
//   id: string
//   nodes: Record<number, Node>
// }

export type ThothReteInput = {
  type: TaskOutputTypes
  outputData: unknown
  task: ThothTask
  key: string
}

export type TaskOutput = {
  type: TaskOutputTypes
  task: ThothTask
  key: string
}

export type ModuleWorkerOutput = WorkerOutputs & {
  [key: string]: any
}

export type ThothWorkerInput = string | unknown | unknown[]
export type ThothWorkerInputs = { [key: string]: ThothWorkerInput[] }
export type ThothWorkerOutputs = WorkerOutputs & {
  [key: string]: TaskOutput
}